Tested a few things with the execution speed.
The main time it takes to execute is from importing the python libraries. If we time after that has all been imported the speed of getting data has gone from 20 ms to 3-4 ms.

Tomorrow I think I will try to make a function which parses an input sentence and recalculates the objective value based on the input (So Random? yes/no, instanceSize 1-1000 or stop) :)