Fixed A lot outside of BFS. BFS takes the longest.
A possible method to speed up BFS:
Replace "j0", "j1" etc with integers. integer lookup is faster than string lookup, which needs to be done very often.

This means that we might not need an unordered_map anymore, but we can use a vector (which is faster.)
Also goes for parent and visited. can be with ints instead of strings.

https://julien.jorge.st/posts/en/effortless-performance-improvements-in-cpp-std-unordered_map/

This details a bit of the performance benchmarks. maybe useful for report.